Adjust some definitions of harassment
Makes slight modifications to some of the definitions of harassment.

Specifically:

 - Explicity prohibits *inappropriate* physical or simulated physical
contact (*high-five*! is not harassment)

 - "Logging online activity" is vague and likely not valid harassment,
especially if the "activity" is publically visible "Accessing private
online activity" is far more precise and constitutes actual wrong doing.

 - Disallow publication of ANY private communication without consent,
regardless of "harassing" nature.  This is to prevent potentially
slanderous statements if one party believes behavior was harassing and
the other does not.  Even if behavior is percieved as harassing, it
should not then be fair game to publically publish, but should instead
be brough privately to the moderators.